% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ace_ai.R
\name{ace_ai}
\alias{ace_ai}
\title{Average Causal Effect with Acceptability Index}
\usage{
ace_ai(cpdag, cov.mat, x, y)
}
\arguments{
\item{cpdag}{The CPDAG with the undefined causal structure}

\item{cov.mat}{The covariance matrix of the set of variables V}

\item{x}{The variable x to determine ICE_x|y-hat and ICE_y|x-hat}

\item{y}{The variable y to determine ICE_x|y-hat and ICE_y|x-hat}
}
\value{
A data frame with information of the computed ICE, the acceptability index and the estimated causal direction.
}
\description{
Computes the Acceptability index as follows:
m(A) = (0.5)*(Al + Ar)
w(A) = (0.5)*(Ar - Al)
Ai = (m(B)-m(A))/(w(B)+w(A))
Sengupta et al. (2000). On comparing interval numbers. European Journal of Operational Research, 127(1)
}
