% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscfunctions.R
\name{mycomparegraphs2}
\alias{mycomparegraphs2}
\title{Function mycomparegraphs2 obtain C(ompleteness) and S(oundness)
#DefinedEdges = DirectedEdges_amEst- DirectedEdges_amCPDAG OR dirgidos de los #UndefinedEdges_amCPDAG 
#TotalEdges = DirectedEdges_amTrue - DirectedEdges_amCPDAG OR #UndefinedEdges_amCPDAG
#CorrectEdges = de los #DefinedEdges aquellos correctos (coinciden con el analogo en el modelo original)
C=#DefinedEdges/#TotalEdges and S=#CorrectEdges/#DefinedEdges}
\usage{
mycomparegraphs2(mdl.true, mdl.est, mdl.cpdag)
}
\description{
Function mycomparegraphs2 obtain C(ompleteness) and S(oundness)
#DefinedEdges = DirectedEdges_amEst- DirectedEdges_amCPDAG OR dirgidos de los #UndefinedEdges_amCPDAG 
#TotalEdges = DirectedEdges_amTrue - DirectedEdges_amCPDAG OR #UndefinedEdges_amCPDAG
#CorrectEdges = de los #DefinedEdges aquellos correctos (coinciden con el analogo en el modelo original)
C=#DefinedEdges/#TotalEdges and S=#CorrectEdges/#DefinedEdges
}
